refactor(credit): modularize ic_quant_engine + credit_scenarios + credit_sensitivity + credit_backtest → quant/ package (Wave 1 PR #6)
Restructure 4 flat files (~1390 LOC) into edgar-style quant/ package:
- models.py: QuantProfile, BacktestInput, CreditBacktestResult, CVStrategy, constants
- parser.py: V1→V2 migration shim + _v2_* field accessors + _parse_* helpers
- profile.py: maturity, rate decomposition, liquidity hooks, risk-adjusted return
- scenarios.py: build_deterministic_scenarios (Base/Downside/Severe)
- sensitivity.py: build_sensitivity_2d + build_sensitivity_3d_summary
- backtest.py: backtest_pd_model (PD/LGD cross-validated)
- service.py: compute_quant_profile orchestrator
- __init__.py: standard imports, __all__ with 9 public exports

Error contract: raises-on-failure (pure computation).
structlog replaces stdlib logging.
All 10 unmapped test callers in test_phase_a_integration.py updated.
