+++ backend/tests/test_vertical_engines.py
@@ -39,7 +39,7 @@ class TestCreditModuleStructure:
     EXPECTED_MODULES = [
         "vertical_engines.credit.deep_review",
         "vertical_engines.credit.critic",
-        "vertical_engines.credit.ic_quant_engine",
+        "vertical_engines.credit.quant",
         "vertical_engines.credit.market_data_engine",
         "vertical_engines.credit.sponsor",
         "vertical_engines.credit.memo_chapter_engine",
